lots of clean up and reafactoring
