+++ src/App.js
@@ -3,13 +3,16 @@ import Paper from '@material-ui/core/Paper';
 import Grid from '@material-ui/core/Grid';
 import { makeStyles } from '@material-ui/core/styles';
 
+import { Typography } from '@material-ui/core';
 import Summary from './components/summary/Summary';
 import SummaryGraph from './components/summary-graph/SummaryGraph';
 import CountryPicker from './components/country-picker/CountryPicker';
 import NavBar from './components/nav-bar/navBar';
 import Layout from './components/layout/Layout';
 import Footer from './components/footer/Footer';
-// import jsonData from './data/mockSummary.json';
+import getSummaryChartFigures from './utils/summary-graph';
+import addGlobalToCountry from './utils/add-global';
+import jsonData from './data/mockSummary.json';
 
 const useStyles = makeStyles((theme) => ({
   root: {
@@ -38,8 +41,8 @@ function App() {
 
   useEffect(() => {
     async function getData() {
-      const response = await fetch('https://api.covid19api.com/summary');
-      const jsonData = await response.json();
+      // const response = await fetch('https://api.covid19api.com/summary');
+      // const jsonData = await response.json();
 
       // Append Global to the list of countries as the first item of the countries array
       const processedAPIData = addGlobalToCountry(jsonData);
@@ -54,61 +57,12 @@ function App() {
 
       setDate(new Date(processedAPIData.Date).toDateString());
 
-      const countriesDataForSummaryFigures = dataForSummaryGraph(
-        processedAPIData.Countries,
-      );
       setSummaryChartFigures(
-        sortDataByConfirmedCases(countriesDataForSummaryFigures),
+        getSummaryChartFigures(processedAPIData.Countries),
       );
     }
     getData();
   }, []);
-
-  function sortDataByConfirmedCases(data) {
-    if (!data.length) {
-      return [];
-    }
-    data.sort((a, b) => a.TotalConfirmed - b.TotalConfirmed);
-
-    return data.reverse().slice(1, 6);
-  }
-
-  function dataForSummaryGraph(data) {
-    return data.map(
-      ({
-        CountryCode,
-        NewConfirmed,
-        TotalConfirmed,
-        NewDeaths,
-        TotalDeaths,
-        NewRecovered,
-        TotalRecovered,
-      }) => {
-        return {
-          CountryCode,
-          NewConfirmed,
-          TotalConfirmed,
-          NewDeaths,
-          TotalDeaths,
-          NewRecovered,
-          TotalRecovered,
-        };
-      },
-    );
-  }
-
-  function addGlobalToCountry(apiResponse) {
-    const newData = {
-      Country: 'Global',
-      Slug: 'global',
-      ...apiResponse.Global,
-    };
-
-    const updatedData = { ...apiResponse };
-    updatedData.Countries.unshift(newData);
-    updatedData.Global = newData;
-    return updatedData;
-  }
 
   function pickCountry(slug) {
     const data = APIData.Countries.filter((c) => c.Slug === slug);
@@ -144,15 +98,17 @@ function App() {
           <Grid item xs={12} />
           <Grid item xs={12}>
             <Paper className={classes.paper}>
-              <Grid container spacing={3}>
-                <Grid item xs={8}>
-                  <Summary figures={figures} date={date} />
-                </Grid>
-                <Grid item xs={4}>
+              <Grid container spacing={5}>
+                <Grid item xs={12} sm={6}>
+                  <Typography variant="h5">Select your Country</Typography>
+                  <br />
                   <CountryPicker
                     pickCountry={pickCountry}
                     countries={APIData.Countries}
                   />
+                </Grid>
+                <Grid item xs={23} sm={6}>
+                  <Summary figures={figures} date={date} />
                 </Grid>
               </Grid>
             </Paper>
