+++ src/components/footer/styles.module.css
@@ -1,17 +1,12 @@
-.list {
-  list-style-type: none;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 0;
-}
-
 .item {
-  float: left;
   margin: 0;
   padding: 0;
   display: flex;
   align-items: center;
+}
+
+.last {
+  justify-content: flex-end;
 }
 
 .icon {
